move mixins from views.py, fix merge by restore autodiscovering rules, fix breadcrumb in issuesubmission
