+++ erudit/apps/userspace/dashboard/views.py
@@ -1,6 +1,6 @@
 from django.views.generic import TemplateView
 
-from core.userspace.viewmixins import LoginRequiredMixin
+from apps.userspace.viewmixins import LoginRequiredMixin
 
 
 class DashboardView(LoginRequiredMixin, TemplateView):
