+++ erudit/apps/userspace/editor/views.py
@@ -11,7 +11,7 @@ from rules.contrib.views import PermissionRequiredMixin
 from navutils import Breadcrumb
 
 from apps.userspace.permissions.views import UserspaceBreadcrumbsMixin
-from core.userspace.viewmixins import LoginRequiredMixin
+from apps.userspace.viewmixins import LoginRequiredMixin
 from core.editor.models import IssueSubmission
 from .forms import IssueSubmissionForm, IssueSubmissionUploadForm
 
@@ -23,7 +23,7 @@ class IssueSubmissionBreadcrumbsMixin(UserspaceBreadcrumbsMixin):
                             self).get_breadcrumbs()
         breadcrumbs.append(Breadcrumb(
             _("Dépôts de numéros"),
-            pattern_name='editor:issues'))
+            pattern_name='userspace:editor:issues'))
         return breadcrumbs
