+++ erudit/apps/userspace/individual_subscription/views.py
@@ -3,36 +3,13 @@ from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
 
 from django_filters.views import FilterView
-from rules.contrib.views import PermissionRequiredMixin
-from navutils import Breadcrumb
 
 from core.individual_subscription.models import IndividualAccount
-from apps.userspace.permissions.views import UserspaceBreadcrumbsMixin
-from core.userspace.viewmixins import LoginRequiredMixin
 
+from .viewmixins import (IndividualAccountBreadcrumbsMixin,
+                         OrganizationCheckMixin)
 from .forms import (IndividualAccountFilter, IndividualAccountForm,
                     IndividualAccountResetPwdForm)
-
-
-class IndividualAccountBreadcrumbsMixin(UserspaceBreadcrumbsMixin):
-
-    def get_breadcrumbs(self):
-        breadcrumbs = super(IndividualAccountBreadcrumbsMixin,
-                            self).get_breadcrumbs()
-        breadcrumbs.append(Breadcrumb(
-            _("Abonnements individuels"),
-            pattern_name='individual_subscription:account_list'))
-        return breadcrumbs
-
-
-class OrganizationCheckMixin(PermissionRequiredMixin, LoginRequiredMixin):
-    permission_required = 'individual_subscription.manage_account'
-
-    def get_queryset(self):
-        qs = IndividualAccount.objects.order_by('-id')
-        ids = [account.id for account in qs if self.request.user.has_perm(
-               'individual_subscription.manage_account', account)]
-        return qs.filter(id__in=ids)
 
 
 class IndividualAccountList(IndividualAccountBreadcrumbsMixin,
