+++ erudit/apps/userspace/journal/viewmixins.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+from django.utils.translation import ugettext_lazy as _
+
+from navutils import Breadcrumb
+
+from apps.userspace.viewmixins import UserspaceBreadcrumbsMixin
+
+
+class JournalBreadcrumbsMixin(UserspaceBreadcrumbsMixin):
+
+    def get_breadcrumbs(self):
+        breadcrumbs = super(JournalBreadcrumbsMixin,
+                            self).get_breadcrumbs()
+        breadcrumbs.append(Breadcrumb(
+            _("Revues"),
+            pattern_name='userspace:journal:journal-information'))
+        return breadcrumbs
