+++ erudit/core/journal/viewmixins.py
@@ -1,23 +1,6 @@
-# -*- coding: utf-8 -*-
-
 from django.http import Http404
-from django.utils.translation import ugettext_lazy as _
-
-from navutils import Breadcrumb
 
 from erudit.models import Journal
-from userspace.views import UserspaceBreadcrumbsMixin
-
-
-class JournalBreadcrumbsMixin(UserspaceBreadcrumbsMixin):
-
-    def get_breadcrumbs(self):
-        breadcrumbs = super(JournalBreadcrumbsMixin,
-                            self).get_breadcrumbs()
-        breadcrumbs.append(Breadcrumb(
-            _("Revues"),
-            pattern_name='journal:journal-information'))
-        return breadcrumbs
 
 
 class JournalCodeDetailMixin(object):
