--- erudit/core/userspace/viewmixins.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from django.contrib.auth.decorators import login_required
-
-
-class LoginRequiredMixin(object):
-    @classmethod
-    def as_view(cls, **initkwargs):
-        view = super().as_view(**initkwargs)
-        return login_required(view)
