+++ erudit/templates/userspace/editor/form.html
@@ -3,10 +3,6 @@
 
 {% block page-title %}{% trans "Ajout d'un numéro" %}{% endblock %}
 
-{% block breadcrumbs %}
-{{ block.super }} / <a href="{% url 'userspace:editor:issues' %}">{% trans "Dépôt de numéros" %}</a> / {% trans "Ajout d'un numéro" %}
-{% endblock %}
-
 {% block extrahead %}
 {{ form.media }}
 {% endblock %}
